pull out opversionoptions
